+++ src/ops.rs
@@ -1,6 +1,30 @@
 extern crate num;
 
 use matrix::Matrix;
+use ops_inplace::VectorVectorOpsInPlace;
+
+// ----------------------------------------------------------------------------
+
+/// Trait for operations on matices.
+pub trait MatrixOps<T> {
+
+    /// Computes the reciprocal (inverse) of each element of the matrix
+    /// and returns the result.
+    fn recip(&self) -> Matrix<T>;
+}
+
+macro_rules! matrix_ops_impl {
+    ($($t:ty)*) => ($(
+
+        impl MatrixOps<$t> for Matrix<$t> {
+            fn recip(&self) -> Matrix<$t> {
+                self.map(|&x| (1.0 as $t) / x)
+            }
+        }
+    )*)
+}
+
+matrix_ops_impl!{ f32 f64 }
 
 
 // ----------------------------------------------------------------------------
@@ -19,7 +43,9 @@ pub trait MatrixScalarOps<T> {
     /// and returns the result.
     fn mul_scalar(&self, scalar: T) -> Matrix<T>;
 
-// TODO div_scalar
+    /// Divides each element of the matrix by a scalar
+    /// and returns the result.
+    fn div_scalar(&self, scalar: T) -> Matrix<T>;
 }
 
 // ----------------------------------------------------------------------------
@@ -46,10 +72,20 @@ macro_rules! matrix_scalar_ops_impl {
                     self.cols()
                 ).unwrap()
             }
+
             fn mul_scalar(&self, scalar: $t) -> Matrix<$t> {
 
                 Matrix::from_vec(
                     self.values().map(|&x| x * scalar).collect(),
+                    self.rows(),
+                    self.cols()
+                ).unwrap()
+            }
+
+            fn div_scalar(&self, scalar: $t) -> Matrix<$t> {
+
+                Matrix::from_vec(
+                    self.values().map(|&x| x / scalar).collect(),
                     self.rows(),
                     self.cols()
                 ).unwrap()
@@ -114,7 +150,11 @@ pub trait VectorVectorOps<T> {
 
     fn sub(&self, rhs: &[T]) -> Vec<T>;
 
-//    fn add(&self, rhs: &[T]) -> Vec<T>;
+    fn add(&self, rhs: &[T]) -> Vec<T>;
+
+    fn mul(&self, rhs: &[T]) -> Vec<T>;
+
+    fn div(&self, rhs: &[T]) -> Vec<T>;
 
     fn mutate<F>(&self, f: F) -> Vec<T>
         where F: Fn(T) -> T;
@@ -128,9 +168,17 @@ macro_rules! vector_vector_ops_impl {
                 self.iter().zip(v.iter()).map(|(&x, &y)| x - y).collect()
             }
 
-//            fn add(&self, v: &[$t]) -> Vec<$t> {
-//                self.iter().zip(v.iter()).map(|(&x, &y)| x + y).collect()
-//            }
+            fn add(&self, v: &[$t]) -> Vec<$t> {
+                self.iter().zip(v.iter()).map(|(&x, &y)| x + y).collect()
+            }
+
+            fn mul(&self, v: &[$t]) -> Vec<$t> {
+                self.iter().zip(v.iter()).map(|(&x, &y)| x * y).collect()
+            }
+
+            fn div(&self, v: &[$t]) -> Vec<$t> {
+                self.iter().zip(v.iter()).map(|(&x, &y)| x / y).collect()
+            }
 
             fn mutate<F>(&self, f: F) -> Vec<$t>
                 where F: Fn($t) -> $t {
@@ -141,7 +189,9 @@ macro_rules! vector_vector_ops_impl {
 
         impl VectorVectorOps<$t> for Vec<$t> {
             fn sub(&self, v: &[$t])                 -> Vec<$t> { (self[..]).sub(v)    }
-//            fn add(&self, v: &[$t])                 -> Vec<$t> { (self[..]).add(v)    }
+            fn add(&self, v: &[$t])                 -> Vec<$t> { (self[..]).add(v)    }
+            fn mul(&self, v: &[$t])                 -> Vec<$t> { (self[..]).mul(v)    }
+            fn div(&self, v: &[$t])                 -> Vec<$t> { (self[..]).div(v)    }
             fn mutate<F: Fn($t) -> $t>(&self, f: F) -> Vec<$t> { (self[..]).mutate(f) }
         }
     )*)
@@ -151,10 +201,14 @@ vector_vector_ops_impl!{ usize u8 u16 u32 u64 isize i8 i16 i32 i64 f32 f64 }
 
 // ----------------------------------------------------------------------------
 
+/// Trait for matrix vector operations.
 pub trait MatrixVectorOps<T> {
 
     /// Adds the given vector to each row of the matrix.
     fn add_row(&self, rhs: &[T]) -> Matrix<T>;
+
+    /// Subtracts the given vector from each row of the matrix.
+    fn sub_row(&self, rhs: &[T]) -> Matrix<T>;
 }
 
 macro_rules! matrix_vector_ops_impl {
@@ -165,15 +219,27 @@ macro_rules! matrix_vector_ops_impl {
             fn add_row(&self, rhs: &[$t]) -> Matrix<$t> {
 
                 let mut m = self.clone();
-                //XXX
-                //mut_row_iter()
+                for i in (0..m.rows()) {
+                    let mut r = m.row_mut(i).unwrap();
+                    r.iadd(rhs);
+                }
+                m
+            }
+
+            fn sub_row(&self, rhs: &[$t]) -> Matrix<$t> {
+
+                let mut m = self.clone();
+                for i in (0..m.rows()) {
+                    let mut r = m.row_mut(i).unwrap();
+                    r.isub(rhs);
+                }
                 m
             }
         }
     )*)
 }
 
-matrix_vector_ops_impl!{ usize u8 u16 u32 u64 isize i8 i16 i32 i64 f32 f64 }
+matrix_vector_ops_impl!{ f32 f64 }
 
 // ----------------------------------------------------------------------------
 
@@ -184,6 +250,16 @@ mod tests {
 
     #[test]
     fn test_matrix_ops() {
+        let m = mat![
+            1.0f32, 2.0; 
+            10.0, 4.0
+        ];
+        let r = m.recip();
+        assert_eq!(r.buf(), &vec![1.0, 0.5, 0.1, 0.25]);
+    }
+
+    #[test]
+    fn test_matrix_scalar_ops() {
 
         let m = mat![
             1.0f32, 2.0; 
@@ -198,6 +274,26 @@ mod tests {
         assert_eq!(b.buf(), &vec![4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0]);
         let c = m.sub_scalar(3.0);
         assert_eq!(c.buf(), &vec![-2.0, -1.0, 0.0, 1.0, 2.0, 3.0, 4.0, 5.0]);
+        let d = m.div_scalar(2.0);
+        assert_eq!(d.buf(), &vec![0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0]);
+    }
+
+    #[test]
+    fn test_matrix_vector_ops() {
+        let m = mat![
+            1.0f32, 2.0; 
+            3.0, 4.0; 
+            5.0, 6.0; 
+            7.0, 8.0
+        ];
+
+        let a = m.add_row(&[2.5, 4.0]);
+        assert_eq!(a.buf(), &vec![3.5, 6.0, 5.5, 8.0, 7.5, 10.0, 9.5, 12.0]);
+
+        let b = m.sub_row(&[4.0, 5.0]);
+        assert_eq!(b.buf(), &vec![-3.0, -3.0, -1.0, -1.0, 1.0, 1.0, 3.0, 3.0]);
+        assert_eq!(m.buf(), &vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0]);
+
     }
 
     #[test]
@@ -221,13 +317,15 @@ mod tests {
     #[test]
     fn test_vector_vector_ops() {
 
-        let a = vec![1.5, 2.0, 3.0, 4.0, 5.0];
+        let a = vec![1.5, 2.0, 2.0, 4.0, 5.0];
         let b = vec![3.0, 2.0, 4.0, 5.0, 1.0];
 
-        assert_eq!(a.sub(&b), vec![-1.5, 0.0, -1.0, -1.0, 4.0]);
-//        assert_eq!(a.add(&b), vec![4.5, 4.0, 7.0, 9.0, 6.0]);
+        assert_eq!(a.sub(&b), vec![-1.5, 0.0, -2.0, -1.0, 4.0]);
+        assert_eq!(a.add(&b), vec![4.5, 4.0, 6.0, 9.0, 6.0]);
+        assert_eq!(a.mul(&b), vec![4.5, 4.0, 8.0, 20.0, 5.0]);
+        assert_eq!(b.div(&a), vec![2.0, 1.0, 2.0, 1.25, 0.2]);
 
-        assert_eq!(a.mutate(|x| x * 2.0), vec![3.0, 4.0, 6.0, 8.0, 10.0]);
+        assert_eq!(a.mutate(|x| x * 2.0), vec![3.0, 4.0, 4.0, 8.0, 10.0]);
     }
 }