+++ src/ops_inplace.rs
@@ -16,10 +16,27 @@ pub trait VectorVectorOpsInPlace<T> {
     ///
     /// let mut v = vec![1.0, 2.0];
     /// let y = vec![3.0, 8.0];
-    /// v.add(&y);
+    /// v.iadd(&y);
     /// assert_eq!(v, vec![4.0, 10.0]);
     /// ```
-    fn add(&mut self, rhs: &[T]);
+    fn iadd(&mut self, rhs: &[T]);
+
+    /// Substracts the given slice `rhs` inplace.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use rustml::ops_inplace::VectorVectorOpsInPlace;
+    ///
+    /// let mut v = vec![1.0, 2.0];
+    /// let y = vec![3.0, 8.0];
+    /// v.isub(&y);
+    /// assert_eq!(v, vec![-2.0, -6.0]);
+    /// ```
+    fn isub(&mut self, rhs: &[T]);
+
+    fn imul(&mut self, rhs: &[T]);
+    fn idiv(&mut self, rhs: &[T]);
 }
 
 macro_rules! impl_vector_vector_ops_inplace {
@@ -27,7 +44,26 @@ macro_rules! impl_vector_vector_ops_inplace {
 
         impl VectorVectorOpsInPlace<$x> for Vec<$x> {
 
-            fn add(&mut self, rhs: &[$x]) {
+            fn iadd(&mut self, rhs: &[$x]) {
+                (self[..]).iadd(rhs);
+            }
+
+            fn isub(&mut self, rhs: &[$x]) {
+                (self[..]).isub(rhs);
+            }
+
+            fn imul(&mut self, rhs: &[$x]) {
+                (self[..]).imul(rhs);
+            }
+
+            fn idiv(&mut self, rhs: &[$x]) {
+                (self[..]).idiv(rhs);
+            }
+        }
+
+        impl VectorVectorOpsInPlace<$x> for [$x] {
+
+            fn iadd(&mut self, rhs: &[$x]) {
 
                 if self.len() != rhs.len() {
                     panic!("Vectors must have the same length.");
@@ -42,6 +78,46 @@ macro_rules! impl_vector_vector_ops_inplace {
                         self.as_ptr() as *mut $z,
                         1             as c_int
                     )
+                }
+            }
+
+            fn isub(&mut self, rhs: &[$x]) {
+
+                if self.len() != rhs.len() {
+                    panic!("Vectors must have the same length.");
+                }
+
+                unsafe {
+                    $y(
+                        self.len()    as c_int,
+                        -1.0          as $z,
+                        rhs.as_ptr()  as *const $z,
+                        1             as c_int,
+                        self.as_ptr() as *mut $z,
+                        1             as c_int
+                    )
+                }
+            }
+
+            fn idiv(&mut self, rhs: &[$x]) {
+
+                if self.len() != rhs.len() {
+                    panic!("Vectors must have the same length.");
+                }
+
+                for (a, b) in self.iter_mut().zip(rhs.iter()) {
+                    *a /= *b;
+                }
+            }
+
+            fn imul(&mut self, rhs: &[$x]) {
+
+                if self.len() != rhs.len() {
+                    panic!("Vectors must have the same length.");
+                }
+
+                for (a, b) in self.iter_mut().zip(rhs.iter()) {
+                    *a *= *b;
                 }
             }
         }
@@ -62,10 +138,14 @@ mod tests {
 
         let mut x: Vec<f32> = vec![1.0, 2.0, 3.0, 4.0];
         let y: Vec<f32> = vec![2.0, 5.0, 9.0, 15.0];
-        x.add(&y);
+        x.iadd(&y);
 
         assert_eq!(x, vec![3.0, 7.0, 12.0, 19.0]);
         assert_eq!(y, vec![2.0, 5.0, 9.0, 15.0]);
+
+        let mut a = [1.0, 2.0, 3.0, 4.0];
+        a.iadd(&y);
+        assert_eq!(a, [3.0, 7.0, 12.0, 19.0]);
     }
 
     #[test]
@@ -73,11 +153,44 @@ mod tests {
 
         let mut x: Vec<f64> = vec![1.0, 2.0, 3.0, 4.0];
         let y: Vec<f64> = vec![2.0, 5.0, 9.0, 15.0];
-        x.add(&y);
+        x.iadd(&y);
 
         assert_eq!(x, vec![3.0, 7.0, 12.0, 19.0]);
         assert_eq!(y, vec![2.0, 5.0, 9.0, 15.0]);
+
+        let mut a = [1.0, 2.0, 3.0, 4.0];
+        a.iadd(&y);
+        assert_eq!(a, [3.0, 7.0, 12.0, 19.0]);
     }
 
+    #[test]
+    fn test_sub_vectorf32() {
+
+        let mut x: Vec<f32> = vec![1.0, 2.0, 3.0, 4.0];
+        let y: Vec<f32> = vec![2.0, 5.0, 9.0, 15.0];
+        x.isub(&y);
+
+        assert_eq!(x, vec![-1.0, -3.0, -6.0, -11.0]);
+        assert_eq!(y, vec![2.0, 5.0, 9.0, 15.0]);
+
+        let mut a = [1.0, 2.0, 3.0, 4.0];
+        a.isub(&y);
+        assert_eq!(a, [-1.0, -3.0, -6.0, -11.0]);
+    }
+
+    #[test]
+    fn test_sub_vectorf64() {
+
+        let mut x: Vec<f64> = vec![1.0, 2.0, 3.0, 4.0];
+        let y: Vec<f64> = vec![2.0, 5.0, 9.0, 15.0];
+        x.isub(&y);
+
+        assert_eq!(x, vec![-1.0, -3.0, -6.0, -11.0]);
+        assert_eq!(y, vec![2.0, 5.0, 9.0, 15.0]);
+
+        let mut a = [1.0, 2.0, 3.0, 4.0];
+        a.isub(&y);
+        assert_eq!(a, [-1.0, -3.0, -6.0, -11.0]);
+    }
 }